Add (de)serialization using serde
The serde (https://serde.rs/) crate is used to implement serialization
and deserialization. The `Deserialize` and `Serialize` traits are
derived using the macros from the serde_derive crate.

I had a need in an application to transmit the Process Unique Id
(PUID) over a TCP stream. I was converting the ProcessUniqueId to
a string, but I thought it would be more elegant and versatile to
implement serialization and deserialization with the serde crate.